car_showroom/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
psycopg2-binary==2.9.3
django-countries==7.2.1
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.2
Markdown==3.3.6
pytz==2021.3
pyuca==1.2
sqlparse==0.4.2
tzdata==2021.5
zipp==3.7.0
django-debug-toolbar





